requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.8.0
bleach==5.0.1
brotlipy==0.7.0
certifi @ file:///C:/b/abs_4f5wo627a3/croots/recipe/certifi_1663615677642/work/certifi
cffi @ file:///C:/Windows/Temp/abs_6808y9x40v/croots/recipe/cffi_1659598653989/work
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography @ file:///C:/ci/cryptography_1652101729823/work
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.1
filelock==3.8.0
huggingface-hub==0.10.0
idna==3.4
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_310/mkl_random_1643050563308/work
mkl-service==2.4.0
mypy-extensions==0.4.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.5.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.12
numpy==1.23.3
oauthlib==3.2.1
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.13.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==304
pywinpty==2.0.8
PyYAML==6.0
pyzmq==24.0.0
qtconsole==5.3.2
QtPy==2.2.0
regex==2022.9.13
requests @ file:///C:/ci/requests_1657735340829/work
requests-oauthlib==1.3.1
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.3.2.post1
stack-data==0.5.0
terminado==0.15.0
textblob==0.17.1
tinycss2==1.1.1
tokenizers==0.12.1
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.22.2
tweepy==4.10.1
typing_extensions @ file:///C:/Windows/TEMP/abs_dd2d0moa85/croots/recipe/typing_extensions_1659638831135/work
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==4.0.3
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
wincertstore==0.2
